Extend plugin to configuration project for java
Add java plugin onto the project so that IDEs can load it as Gradle
project. The configuration is updated to set the Mendix Java sources
folder and the Mendix runtime APIs are added to the project. The
version used for these dependencies is the mendixVersion set on the
extension configration. Additionally the JARs present in the projects
userlib and vendorlib folder are added as dependency.