Improve plugin use when no project is initialized
In case the plugin was added to build.gradle but no project was
initialized yet, Gradle was not usable. This was caused by the
way tasks settins relied on at least `mendixVersion` to be set.

This change moves all configuration to providers which are evaluated
only when their values are needed. With this change the plugin can
add all its tasks to Gradle, accessible for the user. When tasks
are executed that need a Mendix project file to be present, they
will error out which is intented.

Tasks like `mxInit` and `mxListVersions` are now accessible. To
init a project a version is needed which can be provided using
the -P argument on the Gradle CLI.