Improve Java location detection for mxbuild - part 2 (#15)
* Improve Java detection for mxbuild

For the `mxbuild` step the Java location was retrieved solely from
the `JAVA_HOME` environment variable. In case where Java is available
on the `PATH` it also forced users to set `JAVA_HOME`. This is
resolved by taking the value from Gradle as fallback.

* Improve Java detection for mxbuild

For the `mxbuild` step the Java location was retrieved solely from
the `JAVA_HOME` environment variable. In case where Java is available
on the `PATH` it also forced users to set `JAVA_HOME`. This is
resolved by taking the value from Gradle as fallback.